baseline/storm/pyproject.toml
[project]
name = "storm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "h5py>=3.15.1",
    "matplotlib>=3.10.8",
    "ninja>=1.13.0",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pesq>=0.0.4",
    "pydub>=0.25.1",
    "pyroomacoustics>=0.9.0",
    "pysoundfile>=0.9.0.post1",
    "pystoi>=0.4.1",
    "pytorch-lightning>=2.6.0",
    "scipy>=1.17.0",
    "sdeint>=0.3.0",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "torch-ema>=0.3",
    "torchaudio>=2.10.0",
    "torchcodec>=0.10.0",
    "tqdm>=4.67.1",
    "wandb>=0.24.0",
]
